+++ src/components/RunningTimeEntry.tsx
@@ -2,7 +2,7 @@ import { List, Icon, Action, ActionPanel, showToast, Toast } from "@raycast/api"
 import dayjs from "dayjs";
 import duration from "dayjs/plugin/duration";
 import { stopTimeEntry, type User } from "../api";
-import { useCurrentTime } from "../hooks/useCurrentTime";
+import { useCurrentTime } from "../hooks";
 import { categoryColors } from "../helpers";
 
 dayjs.extend(duration);
