+++ src/hooks/useDailyActivities.ts
@@ -4,9 +4,15 @@ import { getDailyActivity } from "../api";
 import type { Dayjs } from "dayjs";
 
 export const useDailyActivities = (dates: Dayjs[]) => {
-  return useCachedPromise(async () => {
+  const { isLoading, data, revalidate } = useCachedPromise(async () => {
     const dailyActivities = await Promise.all(dates.map((date) => getDailyActivity(date)));
 
     return dailyActivities;
   });
+
+  return {
+    isLoadingDailyActivities: isLoading,
+    dailyActivities: data,
+    revalidateDailyActivities: revalidate,
+  };
 };
