+++ src/hooks/useUser.ts
@@ -2,5 +2,11 @@ import { getUser } from "../api";
 import { useCachedPromise } from "@raycast/utils";
 
 export const useUser = () => {
-  return useCachedPromise(getUser);
+  const { isLoading, data, revalidate } = useCachedPromise(getUser);
+
+  return {
+    isLoadingUser: isLoading,
+    user: data,
+    revalidateUser: revalidate,
+  };
 };
